move reset_seed callback
