+++ grid3dfigure.py
@@ -10,10 +10,8 @@ from cluster_sim.app.utils import (
     nx_to_plot,
 )
 import json
-from textwrap import dedent as d
 import dash
-from dash import dcc, html
-from dash.dependencies import Input, Output, State
+from dash import html, Input, Output, State
 import time
 import random
 import jsonpickle
@@ -110,13 +108,6 @@ def initial_call(dummy):
     D = Holes(s.shape)
 
     return jsonpickle.encode(s), G.encode(), D.encode()
-
-
-@app.callback(
-    Output("hover-data", "children"), [Input("basic-interactions", "hoverData")]
-)
-def display_hover_data(hoverData):
-    return json.dumps(hoverData, indent=2)
 
 
 @app.callback(
@@ -250,55 +241,6 @@ def reset_grid(input, xslider, yslider, zslider):
         G.encode(),
         D.encode(),
     )
-
-
-@app.callback(
-    Output("click-data", "children", allow_duplicate=True),
-    Output("draw-plot", "data", allow_duplicate=True),
-    Output("ui", "children", allow_duplicate=True),
-    Output("browser-data", "data", allow_duplicate=True),
-    Output("graph-data", "data", allow_duplicate=True),
-    Output("holes-data", "data", allow_duplicate=True),
-    Input("reset-seed", "n_clicks"),
-    State("load-graph-seed", "value"),
-    State("prob", "value"),
-    State("browser-data", "data"),
-    State("graph-data", "data"),
-    prevent_initial_call=True,
-)
-def reset_seed(nclicks, seed_input, prob, browser_data, graphData):
-    """
-    Randomly measure qubits.
-    """
-    s = jsonpickle.decode(browser_data)
-    G = Grid(s.shape, json=graphData)
-    s.p = prob
-    D = Holes(s.shape)
-    if seed_input:
-        # The user has inputted a seed
-        random.seed(int(seed_input))
-        print(f"Loaded seed : {seed_input}, p = {s.p}")
-        ui = "Loaded seed : {}, p = {}".format(seed_input, s.p)
-    else:
-        # Use a random seed.
-        random.seed()
-        print(f"Loaded seed : {s.seed}, p = {s.p}")
-        ui = "Loaded seed : None, p = {}, shape = {}".format(s.p, s.shape)
-    # p is the probability of losing a qubit
-
-    measurementChoice = "Z"
-
-    for i in range(s.xmax * s.ymax * s.zmax):
-        if random.random() < s.p:
-            if s.removed_nodes[i] == False:
-                s.removed_nodes[i] = True
-                G.handle_measurements(i, measurementChoice)
-                s.log.append(f"{i}, {measurementChoice}; ")
-                s.log.append(html.Br())
-                s.move_list.append([i, measurementChoice])
-                D.add_node(i)
-    D.add_edges()
-    return s.log, 1, ui, jsonpickle.encode(s), G.encode(), D.encode()
 
 
 @app.callback(
